Fix Folly CodeQL warnings (#13263)
* delete unnecessary folly forks

* apply codeql patches

* Change files

* Revert "delete unnecessary folly forks"

This reverts commit d77a1f70d1dba961788bf4090bfb9b7f099173c1.

+++ vnext/Folly/TEMP_UntilFollyUpdate/lang/ToAscii.cpp
@@ -26,7 +26,8 @@ struct to_ascii_array {
   static constexpr data_type_ data_() {
     data_type_ result{};
     Alphabet alpha;
-    for (size_t i = 0; i < Base; ++i) {
+    // [Windows #12703 - CodeQL patch]
+    for (uint64_t i = 0; i < Base; ++i) {
       result.data[i] = alpha(static_cast<uint8_t>(i));
     }
     return result;
@@ -63,7 +64,8 @@ struct to_ascii_table {
   static constexpr data_type_ data_() {
     data_type_ result{};
     Alphabet alpha;
-    for (size_t i = 0; i < Base * Base; ++i) {
+    // [Windows #12703 - CodeQL patch]
+    for (uint64_t i = 0; i < Base * Base; ++i) {
       result.data[i] = //
           (alpha(uint8_t(i / Base)) << (kIsLittleEndian ? 0 : 8)) |
           (alpha(uint8_t(i % Base)) << (kIsLittleEndian ? 8 : 0));
